Added "ippPort()" call to the beginning of the program, to initialize libcups
The CUPS backend runs into a bug of libcups, of the default server
name, port, domain socket file name not being initialized depending on
which functions of the library are used. See this upstream bug report
on CUPS:

https://github.com/apple/cups/issues/5642

A simple call of "ippPort()" in the very beginning of the backend executable
solves the problem.

This also caused the issue

https://github.com/OpenPrinting/cpdb-backend-cups/issues/2

which is solved now, too.