Added test script for "make test"/"make check"
The script src/run-tests.sh first copies all components of the CUPS
installed on the system to run an own instance of CUPS, independnet of
the system and especially outside the system's AppArmor harness.

Then it starts the CUPS backend using this CUPS instance and the
cpdb-text-frontend text mode example frontend of cpdb-libs in one
`dbus-run-session` environment. The frontend discovers the backend and
some test tasks are performed with it, like listing options and
printing a job.

Afterwards the frontend is stopped by supplying "stop" to its standard
input. If the frontend does not stop, it gets killed after a
timeout. The log of the frontend is checked whether it got actually
stopped by the "stop" command or not.

The backend is shut down by a TERM signal. After a short delay any
remaining sub-process is killed via KILL.

The test fails if the backend is not discovered, one of the tasks
failed, or if a sub-process needs to be KILLed.

This test script requires CUPS, the libraries of cpdb-libs, and the
packages cpdb-libs-tools and dbus-run-session (usually part of the
dbus-daemon package) to be installed.