Build system: Removed unnecessary lines in tools/Makefile.am
Removed the TESTdir and TEST_SCRIPTS entries in tools/Makefile.am.
They are not needed and let `make install` try to install
`run-tests.sh` in the source directory, where it already is, causing
an error.

The problem only occurs when `make install` is run without
`DESTRDIR=...`. Then the installation runs without errors.