get_all_media(): Do not crash on custom page size range entries
The media-col-database IPP attribute contains one entry for each valid
combination of page size dimensions, margins, and in some cases also
media source and media type.

If custom page sizes are supported, this is represented by one
media-col-database entry (usually the last one) where the width and
length dimensions are ranges, consisting of 2 numbers each. The
get_all_media() tried to read those as normal intergers, getting back
0 and this cause NULL being returned by pwgMediaForSize() function and
a segfault when trying to get the PWG name of the page size.

This is now fixed by skipping the entry for the custom page size
(actually we need to support custom page sizes).